Changes to be committed: 	modified:   src/app/api/comments/[id]/route.ts 	modified:   src/app/bookmarks/page.tsx 	modified:   src/app/profile/[username]/[id]/ProfileLinkMenu.tsx 	new file:   src/app/profile/[username]/[id]/followers/FollowersUsers.tsx 	modified:   src/app/profile/[username]/[id]/followers/page.tsx 	new file:   src/app/profile/[username]/[id]/following/FollowingUsers.tsx 	modified:   src/app/profile/[username]/[id]/following/page.tsx 	modified:   src/app/profile/[username]/[id]/layout.tsx 	modified:   src/app/profile/[username]/[id]/likes/page.tsx 	modified:   src/components/CommentModal.tsx 	modified:   src/components/FollowButton.tsx 	modified:   src/components/Modal.tsx 	modified:   src/components/PostBlock.tsx 	modified:   src/components/Sidebar.tsx 	modified:   src/components/ToggleMode.tsx 	modified:   src/components/UserMenu.tsx 	new file:   src/components/UserSessionMenu.tsx 	renamed:    src/components/UsersList.tsx -> src/components/UsersWidget.tsx 	modified:   src/components/Widgets.tsx

+++ src/app/api/comments/[id]/route.ts
@@ -1,11 +1,14 @@
 import { prisma } from '@/lib/db/prisma';
-import { stat } from 'fs';
 import { NextResponse } from 'next/server';
 
 export const DELETE = async (
   req: Request,
   { params }: { params: { id: string } }
 ) => {
+  if (!params.id) {
+    console.log('Missing requirements fields');
+    return NextResponse.json({ error: 'Missing requirements fields ' });
+  }
   try {
     const comment = await prisma.comment.delete({
       where: { id: params.id },
@@ -17,25 +20,23 @@ export const DELETE = async (
   }
 };
 
-const getCurrentDateTime = () => {
-  const now = new Date();
-  return now.toISOString(); // Повертаємо дату та час у форматі ISO, наприклад: "2023-07-18T12:34:56.789Z"
-};
-
 export const PUT = async (
   req: Request,
   { params }: { params: { id: string } }
 ) => {
   const body = await req.json();
-
-try {
-  const comment = await prisma.comment.update({
-    where: { id: params.id },
-    data: { content: body },
-  });
-  return new NextResponse(JSON.stringify(comment), {status: 201})
-} catch (error) {
-  console.log(error)
-  return new NextResponse('DATABASE-ERROR ', {status: 500})
-}
+  if (!params.id || !body) {
+    console.log('Missing requirements fields');
+    return NextResponse.json({ error: 'Missing requirements fields ' });
+  }
+  try {
+    const comment = await prisma.comment.update({
+      where: { id: params.id },
+      data: { content: body },
+    });
+    return new NextResponse(JSON.stringify(comment), { status: 201 });
+  } catch (error) {
+    console.log(error);
+    return new NextResponse('DATABASE-ERROR ', { status: 500 });
+  }
 };
